Relacion categories user admin

+++ e-commer/src/main/java/com/lazxnet/e_commer/categories/Service/CategoryService.java
@@ -1,7 +1,11 @@
 package com.lazxnet.e_commer.categories.Service;
 
+import com.lazxnet.e_commer.categories.DTO.CategoryResponseDTO;
 import com.lazxnet.e_commer.categories.Entity.Category;
 import com.lazxnet.e_commer.categories.Repository.CategoryRepository;
+import com.lazxnet.e_commer.userAdmin.Entity.UserAdmin;
+import com.lazxnet.e_commer.userAdmin.Repository.UserAdminRepository;
+import com.lazxnet.e_commer.userAdmin.dto.UserAdminResponseDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,20 +19,59 @@ public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
 
-    //Crear Categoria
-    public  Category createCategory(Category category) {
+    @Autowired
+    private UserAdminRepository userAdminRepository;
 
-        return categoryRepository.save(category);
+    //Crear Categoria
+    public CategoryResponseDTO createCategory(Category category, UUID userAdminId) {
+
+        //TODO: Validar que el UserAdminId exista
+        UserAdmin userAdmin = userAdminRepository.findById(userAdminId)
+                .orElseThrow(()-> new RuntimeException("UserAdmin no encontrado"));
+
+        category.setUserAdmin(userAdmin);
+        Category savedCategory = categoryRepository.save(category);
+        return convertToResponseDTO(savedCategory);
+    }
+
+
+    //Metodo para convectir entidad a DTO
+    private CategoryResponseDTO convertToResponseDTO(Category category) {
+        CategoryResponseDTO dto = new CategoryResponseDTO();
+        dto.setCategoryId(category.getCategoryId());
+        dto.setName(category.getName());
+        dto.setDescription(category.getDescription());
+
+        //Convertir UserAdmin a UserAdminResponseDTO
+        UserAdmin userAdmin = category.getUserAdmin();
+        UserAdminResponseDTO userAdminDTO = new UserAdminResponseDTO();
+        userAdminDTO.setUserAdminId(userAdmin.getUserAdminId());
+        userAdminDTO.setEmail(userAdmin.getEmail());
+        userAdminDTO.setFullName(userAdmin.getFullName());
+
+        dto.setUserAdmin(userAdminDTO);
+        return dto;
     }
 
     //Obtener todas las Categorias
-    public List<Category> getAllCategories() {
-        return categoryRepository.findAll();
+    public List<CategoryResponseDTO> getAllCategories() {
+        return categoryRepository.findAll()
+        .stream()
+        .map(this :: convertToResponseDTO)
+        .collect(java.util.stream.Collectors.toList());
     }
 
     //Borrar categoria por id
-    public Category deleteCategoryById(UUID id) {
-        Category category = categoryRepository.findById(id).orElse(null);
+    public Category deleteCategoryById(UUID id, UUID userAdminId) {
+
+        if (!userAdminRepository.existsById(userAdminId)) {
+            throw new RuntimeException("UserAdmin no encontrado");
+        }
+
+        Category category = categoryRepository
+                .findById(id)
+                .orElseThrow(()-> new RuntimeException("Categoria no encontrada"));
+
         categoryRepository.delete(category);
         return category;
     }
